Add unit tests for REdux additions
