+++ api-catalog-ui/frontend/package.json
@@ -80,7 +80,7 @@
         "cors": "2.8.5",
         "cross-env": "5.2.0",
         "cypress": "6.5.0",
-        "cypress-file-upload": "^5.0.8",
+        "cypress-file-upload": "5.0.8",
         "enzyme": "3.7.0",
         "enzyme-adapter-react-16": "1.15.6",
         "eslint-config-airbnb": "16.1.0",
